Allow passing gct as a function paramater in the GC
This patch adds support for passing the gct variable as a
paramter in the garbage collector. The GC functions that
need access to the gct variable are declared with the
DECLARE_GCT_PARAM macro that will add the gct as an extra
argument. When calling these function use the GCT_PARAM
macro to automatically add the gct variable to the function
call.

The defintion of these two macros are controlled by the
PASS_GCT_AS_PARAM preprocessor variable. If this variable is
defined then the above macros are setup to do the parameter
passing, otherwise these macros will be suitably defined to
do nothing.

Currently we only pass the gct variable as a paramter when
using an llvm compiler C compiler such as clang or llvm-gcc.