Ignore stdcall c-call in native codegen on x86_64
The stdcall calling convention is not supported on x86_64.
When an ffi import requests stdcall, a warning is issued as
desired by #3336. However, the native codegen was still
generating code that expected the callee to cleanup the
stack arguments when calling a c function that requests
stdcall.

This patch changes the codegen to actually use the ccall
calling convention as intended.