Remove RTS_PRIVATE qualification on some rts functions
The RTS_PRIVATE qualification issues instructions to GCC to
not have the function be externally visible in a dynamic
library. This was causing problems with the lli interpreter
which was trying to reference these functions through the
dynamic linker. Allowing these functions to be visible is
necessary so that we can collect profiling data through an
lli run.