Have the rts use llvm when compiling .cmm files
We want the rts to use llvm to compile the cmm files instead
of going through the native code generator. We need to use
llvm so that we can fairly compare the performance of a non
trace-based program with a trace-based program since the
trace-based program will directly use the llvm bitcode from
the cmm files.

Without this change the non-trace-based programs had an
unfair advantage since the native codegen compiles
Updates.cmm better than llvm (see #4308).